Add method for more specific instruction for writing to file
Determine what file signature the byte code has. If it's a mp4 file
write the data to a file without any modifications of the byte code.
If it's a jpeg file replace the IPTC special instructions with zeros,
because it could be used for tracking the usage of the downloaded file.