Create HTTPHeaders class
Rewrite random_useragent() to be a method of HTTPHeaders.

The headers with a random user agent is set when the HTTPHeader class is initialised
with a file to a list of user agents.

To get a new user agent just reassign the list of user agents to the objects
headers attribute.