Refactor _save() method in Downloader class
Removed the _write() method and moved the file signature conditions
to the _save() method instead.

If the bytes in the buffer contains a JPEG signature, save it with
Pillow (PIL) to remove any unwanted meta data within the file. Also
try to keep the quality of the original JPEG file.

If the bytes on the other hand contains a MP¤ signature just save
it with the builtin open() function.