Add limit checker for post url scraper
Before the limit was checked there was a bug that either crashed the
program or made it run forever without any urls being scraped.

The limit checker fixes this by making sure the limit for how many posts
that will be downloaded is less than or equal to the users or hashtags
total number of uploaded posts.