Add new feature to --output flag
-o or --output now requires two arguments. 1) Path to a directory on the file system
and 2) a custom directory name for the folder that will store all downloaded files.

Before only a path was required and the default name for the output directory in that
path was always "downloads". That was limiting because you would have to make sure not
to give a path where a downloads file already existed. Now multiple output directories
can share the same path.