Add setup.py and update CLI
The code can now run as a CLI program without first calling python.
To do that you have to install the package with "pip" and then just
type the programs name in the terminal to run it.

The package can also be imported and used in other python projects now.